requirements.txt
streamlit
opencv-python-headless
openai
Pillow 